python-rag-agent-project/requirements.txt
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
pypdf
python-dotenv